Disable Heap Malloc on Linux
Disables heap malloc calls on Linux, as there are games where the heap malloc causes a crash.  Note that the assert had to be modified, as pthread_gethandle is a winpthreads specific method.

Also fixes a minor oversight that caused games not to perform TLS mallocs on Windows when there was no heap api.