Automatically Update shadPS4 on OS Startup (#2700)
Created a separate .exe for updating. Added a checkbox in the settings
to enable the option to update on Windows startup. When the option is
enabled, a PowerShell script adds the new updater exe to the Windows
Task Scheduler and sets it to run at startup. If there are updates
available at startup, the update windows pops up. Disabling the option
runs another script to remove the updater from the Task Scheduler.

Co-authored-by: Joao Ribeiro <joao.viegas.ribeiro@tecnico.ulisboa.pt>